requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
cffi==1.17.1
discord==2.3.2
discord.py==2.5.2
dotenv==0.9.9
frozenlist==1.7.0
idna==3.10
multidict==6.6.3
propcache==0.3.2
pycparser==2.22
PyNaCl==1.5.0
python-dotenv==1.1.1
typing_extensions==4.14.1
yarl==1.20.1
yt-dlp==2025.7.21
